Refactor `process_transaction` remments
