+++ ledger_bot/message_generators/__init__.py
@@ -4,6 +4,7 @@ from .generate_list_message import generate_list_message
 from .generate_reminder_status_message import generate_reminder_status_message
 from .generate_stats_message import generate_stats_message
 from .generate_transaction_status_message import generate_transaction_status_message
+from .send_message import send_message
 
 __all__ = [
     "generate_help_message",
@@ -11,4 +12,5 @@ __all__ = [
     "generate_reminder_status_message",
     "generate_stats_message",
     "generate_transaction_status_message",
+    "send_message",
 ]
