--- ledger_bot/process_transactions/__init__.py
@@ -1,10 +1,0 @@
-"""Functions to update transaction statuses."""
-
-
-from .refresh_transaction import refresh_transaction
-from .send_message import send_message
-
-__all__ = [
-    "refresh_transaction",
-    "send_message",
-]
